+++ Parse/src/test/java/com/parse/PushServiceTest.java
@@ -66,17 +66,6 @@ public class PushServiceTest extends ResetPluginsParseTest {
   }
 
   @Test
-  public void testDefaultHandler() {
-    ManifestInfo.setPushType(PushType.NONE);
-    PushHandler handler = PushServiceUtils.createPushHandler();
-    assertTrue(handler instanceof PushHandler.FallbackHandler);
-
-    ManifestInfo.setPushType(PushType.GCM);
-    handler = PushServiceUtils.createPushHandler();
-    assertTrue(handler instanceof GcmPushHandler);
-  }
-
-  @Test
   public void testOnCreateWithoutInit() {
     ParsePlugins.reset();
     controller.create();
